Remove vgsshow-pdf module, move VGSPDFView.kt to vgsshow module, update comments
