--- vgsshow-pdf/src/main/AndroidManifest.xml
@@ -1,18 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<manifest xmlns:android="http://schemas.android.com/apk/res/android"
-    package="com.verygoodsecurity.vgsshow.pdf">
-
-    <application>
-
-        <provider
-            android:name="androidx.core.content.FileProvider"
-            android:authorities="com.verygoodsecurity.vgsshow.pdf.provider"
-            android:exported="false"
-            android:grantUriPermissions="true">
-
-            <meta-data
-                android:name="android.support.FILE_PROVIDER_PATHS"
-                android:resource="@xml/file_paths" />
-        </provider>
-    </application>
-</manifest>
