+++ vgsshow/build.gradle
@@ -49,6 +49,10 @@ dependencies {
     // Http
     api 'com.squareup.okhttp3:okhttp:4.9.0'
 
+    // PDF viewer.
+    // NOTE: Used by VGSPDFView. Consumer should add this dependency by itself as it has big size.
+    compileOnly 'com.github.barteksc:android-pdf-viewer:2.8.2'
+
     testImplementation 'junit:junit:4.13.1'
     androidTestImplementation 'androidx.test.ext:junit:1.1.2'
     androidTestImplementation 'androidx.test.espresso:espresso-core:3.3.0'
