+++ vgsshow/src/main/java/com/verygoodsecurity/vgsshow/core/helper/ViewsStore.kt
@@ -6,8 +6,8 @@ import com.verygoodsecurity.vgsshow.core.network.model.data.response.ResponseDat
 import com.verygoodsecurity.vgsshow.util.extension.logStartViewsUpdate
 import com.verygoodsecurity.vgsshow.util.extension.logWaring
 import com.verygoodsecurity.vgsshow.widget.VGSTextView
-import com.verygoodsecurity.vgsshow.widget.core.VGSRenderView
 import com.verygoodsecurity.vgsshow.widget.core.VGSView
+import com.verygoodsecurity.vgsshow.widget.view.pdf.VGSPDFView
 
 internal class ViewsStore {
 
@@ -40,7 +40,7 @@ internal class ViewsStore {
             } else {
                 when (view) {
                     is VGSTextView -> view.setText(revealedData)
-                    is VGSRenderView<*> -> view.render(Base64.decode(revealedData, Base64.NO_WRAP))
+                    is VGSPDFView -> view.render(Base64.decode(revealedData, Base64.NO_WRAP))
                     else -> throw IllegalArgumentException("Not implemented yet!")
                 }
             }
