+++ vgsshow/src/main/java/com/verygoodsecurity/vgsshow/widget/extension/Context.kt
@@ -7,7 +7,7 @@ import android.content.res.TypedArray
 import android.util.AttributeSet
 import androidx.core.content.res.use
 
-fun Context.getStyledAttributes(
+internal fun Context.getStyledAttributes(
     attributeSet: AttributeSet?,
     styleArray: IntArray, block: TypedArray.() -> Unit
 ) = this.obtainStyledAttributes(attributeSet, styleArray).use(block)
