+++ vgsshow/src/main/res/values/attrs.xml
@@ -179,4 +179,25 @@
         <attr name="secureTextEnd"/>
         <attr name="secureTextSymbol"/>
     </declare-styleable>
+
+    <declare-styleable name="VGSPDFView">
+
+        <!-- Default start page.-->
+        <attr name="defaultPage" format="integer" />
+
+        <!-- Enable/Disable changing pages using scrolling.-->
+        <attr name="isSwipeEnabled" format="boolean" />
+
+        <!-- Enable/Disable horizontal scrolling.-->
+        <attr name="isSwipeHorizontalEnabled" format="boolean" />
+
+        <!-- Enable/Disable double tap.-->
+        <attr name="isDoubleTapEnabled" format="boolean" />
+
+        <!-- Improve rendering on low-res screens.-->
+        <attr name="isAntialiasingEnabled" format="boolean" />
+
+        <!-- Spacing between pages in dp. -->
+        <attr name="spacing" format="integer" />
+    </declare-styleable>
 </resources>
