+++ vgsshow/src/main/res/xml/file_paths.xml
